backend/bot/handlers/audio_handler.py
from linebot import LineBotApi
from linebot.models import AudioSendMessage, TextSendMessage
from django.conf import settings

line_bot_api = LineBotApi(settings.LINE_ACCESS_TOKEN)

def process_audio_message(event):
    """
    Handle incoming audio messages
    """
    try:
        # Get the audio content
        message_content = line_bot_api.get_message_content(event.message.id)
        
        # For now, just acknowledge receipt of audio
        line_bot_api.reply_message(
            event.reply_token,
            TextSendMessage(text="I received your audio message!")
        )
        
        # TODO: Add your audio processing logic here
        # You might want to:
        # 1. Save the audio file
        # 2. Process it (speech-to-text, analysis, etc.)
        # 3. Send appropriate response
        
    except Exception as e:
        line_bot_api.reply_message(
            event.reply_token,
            TextSendMessage(text=f"Sorry, I couldn't process your audio message.")
        ) 