backend/requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Deprecated==1.2.15
Django==5.1.3
dnspython==2.7.0
frozenlist==1.5.0
future==1.0.0
idna==3.10
line-bot-sdk==3.14.2
Markdown==3.7
multidict==6.1.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
u==0.1
urllib3==2.2.3
wrapt==1.16.0
yarl==1.17.1
gunicorn
line-bot-sdk==3.14.2
python-decouple==3.8
mysqlclient==2.2.6